Added buttons to the game, I need some news assets. The bottom of the window will be used for tool tips and the top of the window is going to be used to show information. The upgrade button of the tree will be located on the tree itself.

+++ src/upgradegrid.h
@@ -5,7 +5,7 @@ class Upgrades{
 public:
 	Upgrades(	std::string defaultImageButtonLocation, std::string hoveredImageButtonLocation,
 				std::string clickedImageButtonLocation, int getWindowWidth, int getWindowHeight,
-				int initialX, int initialY, int gap, SDL_Renderer *getRenderer);
+				float initialX, float initialY, float gap, SDL_Renderer *getRenderer);
 				
 	~Upgrades();
 	void ShowUpgradeButtons(int _x, int _y, bool _isClicked);
